feat: Add Images Endpoint Aggregation in users, publications, tags

+++ controllers/publicationsImages.controller.js
@@ -13,37 +13,48 @@ const imagesPublicationsService = new ImagesPublicationsService()
 
 
 const uploadImagePublication = async (request, response, next) => {
-  const { id } = request.params
-  const files = request.files
+  const { id: idPublication } = request.params
+  const files = request.files;
+  const { isAdmin, isSameUser } = request.user;
+
   try {
 
-    if (files.length) {
-      let imagesKeys = []
-      await imagesPublicationsService.publicationImagesExist(id)
-      let order = 0
-      await Promise.all(files.map(async (file) => {
+    if (isAdmin || isSameUser) {
 
-        const imageName = uuid.v4()
-        const fileResize = await sharp(file.path)
-          .resize({ height: 1920, width: 1080, fit: 'contain' })
-          .toBuffer()
+      if (files.length) {
+        let imagesKeys = []
+        await imagesPublicationsService.publicationImagesExist(idPublication)
+        await imagesPublicationsService.canUploadImages(idPublication);
+        response.json({ message: 'esta' })
 
-        await uploadFile(fileResize, imageName, file.mimetype)
+        let order = 1
+        await Promise.all(files.map(async (file) => {
 
-        let newImagePublication = await imagesPublicationsService.createImage(imageName, id, order)
-        order++
-        imagesKeys.push(newImagePublication.image_url)
-      }))
-      await Promise.all(files.map(async (file) => {
-        await unlinkFile(file.path)
-      }))
-      return response
-        .status(200)
-        .json({ results: { message: 'success upload', images: imagesKeys } });
-    } else {
-      throw new CustomError('Images were not received', 404, 'Not Found')
+          const idImage = uuid.v4()
+          let imageName = `publications-images-${idPublication}-${idImage}`
+          const fileResize = await sharp(file.path)
+            .resize({ height: 1920, width: 1080, fit: 'contain' })
+            .toBuffer()
+
+          await uploadFile(fileResize, imageName, file.mimetype)
+
+          let newImagePublication = await imagesPublicationsService.createImage(imageName, idPublication, order)
+          order++
+          imagesKeys.push(newImagePublication.image_url)
+        }))
+        await Promise.all(files.map(async (file) => {
+          await unlinkFile(file.path)
+        }))
+        return response
+          .status(200)
+          .json({ results: { message: 'success upload', images: imagesKeys } });
+
+      } else {
+        throw new CustomError('Images were not received', 404, 'Not Found')
+      }
     }
 
+    throw new CustomError('You are not authorized to make changes to this user', 403, 'Unauthorized');
   } catch (error) {
     if (files) {
       await Promise.all(files.map(async (file) => {
@@ -52,15 +63,20 @@ const uploadImagePublication = async (request, response, next) => {
     }
     next(error)
   }
+
 }
 
 const destroyImageByPublication = async (request, response, next) => {
   try {
-    const { idImage } = request.params;
-    let imagePublication = await imagesPublicationsService.getImageOr404(idImage)
-    await deleteFile(imagePublication.image_url)
-    await imagesPublicationsService.removeImage(idImage)
-    return response.status(200).json({ message: 'Image Deleted', idPublication: imagePublication.publication_id, idImage: idImage })
+    const { isAdmin, isSameUser } = request.user;
+    const { order, id } = request.params;
+    if (isAdmin || isSameUser) {
+      let imagePublication = await imagesPublicationsService.getImageOr404(order, id)
+      await deleteFile(imagePublication.image_url)
+      await imagesPublicationsService.removeImage(imagePublication.id)
+      return response.status(200).json({ message: 'Image Deleted', idPublication: imagePublication.publication_id, image_order: imagePublication.order })
+    }
+    throw new CustomError('You are not authorized to make changes to this user', 403, 'Unauthorized');
   } catch (error) {
     next(error)
   }
@@ -118,10 +134,14 @@ const updateImage = async (request, response, next) => {
   try {
     const body = request.body;
     const { id } = request.params;
-    const result = await imagesPublicationsService.updateOrderImages(body, id);
-    response.json({
-      message: 'has been successfully updated'
-    })
+    const { isSameUser } = request.user;
+    if (isSameUser) {
+      const result = await imagesPublicationsService.updateOrderImages(body, id);
+      return response.json({
+        message: 'has been successfully updated'
+      })
+    }
+    throw new CustomError('You are not authorized to make changes to this user', 403, 'Unauthorized');
   } catch (error) {
     next(error)
   }
